requirements.txt
selenium
python-dotenv
Flask
waitress